+++ mkdoxy/finder.py
@@ -2,7 +2,7 @@ from typing import Dict
 
 from mkdoxy.constants import Kind
 from mkdoxy.doxygen import Doxygen
-from mkdoxy.utils import recursive_find, recursive_find_with_parent
+from mkdoxy.node import Node
 
 
 class Finder:
@@ -10,13 +10,15 @@ class Finder:
         self.doxygen = doxygen
         self.debug = debug
 
-    def _normalize(self, name: str) -> str:
+    @staticmethod
+    def _normalize(name: str) -> str:
         return name.replace(" ", "")
 
-    def list_to_names(self, list):
+    @staticmethod
+    def list_to_names(list):
         return [part.name_params for part in list]
 
-    def doxyClass(self, project, className: str):
+    def doxy_class(self, project, className: str):
         classes = recursive_find(self.doxygen[project].root.children, Kind.CLASS)
         if classes:
             for findClass in classes:
@@ -26,7 +28,7 @@ class Finder:
         return None
 
     def doxy_class_method(self, project, className: str, methodName: str):
-        findClass = self.doxyClass(project, className)
+        findClass = self.doxy_class(project, className)
         if findClass:
             if isinstance(findClass, list):
                 for member in findClass:
@@ -60,3 +62,23 @@ class Finder:
                     return file
             return self.list_to_names(files)
         return None
+
+
+def recursive_find(nodes: list["Node"], kind: Kind) -> list["Node"]:
+    ret = []
+    for node in nodes:
+        if node.kind == kind:
+            ret.append(node)
+        if node.kind.is_parent():
+            ret.extend(recursive_find(node.children, kind))
+    return ret
+
+
+def recursive_find_with_parent(nodes: list["Node"], kinds: list[Kind], parent_kinds: list[Kind]) -> list["Node"]:
+    ret: list["Node"] = []
+    for node in nodes:
+        if node.kind in kinds and node.parent is not None and node.parent.kind in parent_kinds:
+            ret.append(node)
+        if node.kind.is_parent() or node.kind.is_dir() or node.kind.is_file():
+            ret.extend(recursive_find_with_parent(node.children, kinds, parent_kinds))
+    return ret
