+++ mkdoxy/generator_base.py
@@ -14,9 +14,8 @@ from mkdoxy.node import DummyNode, Node
 from mkdoxy.utils import (
     merge_two_dicts,
     load_template_meta,
-    recursive_find,
-    recursive_find_with_parent,
 )
+from mkdoxy.finder import recursive_find, recursive_find_with_parent
 
 log: logging.Logger = logging.getLogger("mkdocs")
