+++ mkdoxy/node.py
@@ -7,7 +7,6 @@ from mkdoxy.constants import OVERLOAD_OPERATORS, Kind, Visibility
 from mkdoxy.markdown import escape
 from mkdoxy.project import ProjectContext
 from mkdoxy.property import Property
-from mkdoxy.utils import split_safe
 from mkdoxy.xml_parser import XmlParser
 
 log: logging.Logger = logging.getLogger("mkdocs")
@@ -522,7 +521,7 @@ class Node:
     def name_tokens(self) -> [str]:
         if self.is_dir or self.is_file:
             return self._name.split("/")
-        return split_safe(self._name, "::")
+        return self._split_safe(self._name, "::")
 
     @property
     def name_short(self) -> str:
@@ -848,6 +847,46 @@ class Node:
             ret += self._print_node_recursive_md(child, depth + 1)
 
         return ret
+
+    @staticmethod
+    def _contains(a, pos, b):
+        ai = pos
+        bi = 0
+        if len(b) > len(a) - ai:
+            return False
+        while bi < len(b):
+            if a[ai] != b[bi]:
+                return False
+            ai += 1
+            bi += 1
+        return True
+
+    def _split_safe(self, s: str, delim: str) -> [str]:
+        tokens = []
+        i = 0
+        last = 0
+        inside = 0
+        while i < len(s):
+            c = s[i]
+            if i == len(s) - 1:
+                tokens.append(s[last : i + 1])
+            if c in ["<", "[", "{", "("]:
+                inside += 1
+                i += 1
+                continue
+            if c in [">", "]", "}", ")"]:
+                inside -= 1
+                i += 1
+                continue
+            if inside > 0:
+                i += 1
+                continue
+            if self._contains(s, i, delim):
+                tokens.append(s[last:i])
+                i += 2
+                last = i
+            i += 1
+        return tokens
 
 
 class DummyNode:
